feat(api): sanitize product description (#2050)
## What's the purpose of this pull request?

This PR adds support for HTML in product descriptions.

## How it works?

FastStore API sanitizes the product description in the backend. This
ensures no possibly vulnerable tags, such as `script` tags are sent and
executed on the frontend.

The frontend now inserts the description as innerHTML, which allows it
to be properly rendered.

## How to test it?

Browse the two deploy previews below. Note on how one of the product
descriptions, the HTML renders as text, and, in the other that uses this
version of the packages it is interpreted as actual HTML.

### Starters Deploy Preview


https://sfj-5d275d1--starter.preview.vtex.app/unbranded-frozen-pizza-1670312/p
https://starter.vtex.app/unbranded-frozen-pizza-1670312/p

## References

https://www.npmjs.com/package/sanitize-html