fix: rollback prettier to 2ˆ to fix publishing issues (#2056)
## What's the purpose of this pull request?

This PR fixes
[this](https://github.com/vtex/faststore/actions/runs/6458890810/job/17533535581)
issue that happened while publishing packages due to incompatible
`eslint-plugin-prettier` versions with prettier 3.x.

## How it works?

In #2050 I upgraded the prettier version to 3.x. That caused the issue
because `tsdx` has a dependency on an old `eslint-plugin-prettier`
version, which is incompatible with `prettier@3`. I'm rolling back the
upgrade and fixing the prettier version on the major `2` which makes it
difficult for others to make the same mistake as me.

In the future, we remove all uses of `tsdx` as it is no longer supported
and has gone 3 years without updates.

## How to test it?

By merging this!

## References

https://github.com/prettier/eslint-plugin-prettier/issues/562
Thanks @hellofanny for noticing the error and investigating it!