chore: runtime imports config (#2071)
## What's the purpose of this pull request?

This PR makes it possible to use [nextjs's recommended
syntax](https://nextjs.org/docs/pages/api-reference/next-config-js/redirects)
for declaring redirects.

It also eliminates the need to use BASE_PATH.

## How it works?

By removing the merge from the CLI and moving it to faststore core, we
remove dependencies from the CLI and avoid future headaches by stopping
to rely on `stringify-object`. There's nothing essentially wrong with
the package, but it's one less point of failure.

`stringify-object` wouldn't parse nextjs's recommended syntax for
redirects, which would led to restrictions on how users could define it
in the browser. This restriction no longer exists.

Since the starter's config file is copied under the same folder as the
`src` folder and the path structure is preserved, there's no need to use
BASE_PATH anymore, which leads to the same API offered by nextjs, which
is awesome.

## How to test it?

Define redirects inside your starter. The only restriction is that any
file imported by the redirects function must be under the `src` folder.
Test in the deploy preview of the starter below by accessing `/my-test`
and be redirected to `/my-test-new`.

### Starters Deploy Preview

https://github.com/vtex-sites/starter.store/pull/232

## References

#2070 
#2069