chore: remove copying faststore config during build (#2074)
## What's the purpose of this pull request?

The faststore.config.js file from the `.faststore` folder was being
copied to the starter after the build, which caused it to throw errors
because it couldn't find default definitions for the config or the
copied starter config present inside `.faststore`.

Lighthouse was not working because of this.

## How it works?

This PR makes the CLI not copy the `.faststore/faststore.config.js` file
to the starter after the build.

## How to test it?

Run `yarn build` on the starter. You shouldn't see any changes in the
starter's `faststore.config.js` file.

### Starters Deploy Preview

https://github.com/vtex-sites/starter.store/pull/232

The integration tests are failing due to an unsupported workflow on
webops. Lighthouse actually started working after this change!