Feat: add XML to list of public filetypes copied during build (#2104)
## What's the purpose of this pull request?

The recent change to support copying the contents of the `/public` dir
during build allows users to customize the `robots.txt` file, but
doesn't enable the users to upload a standard XML sitemap which the
`robots.txt` file can link to.

## How it works?

Adds `xml` to the list of filetypes.

## How to test it?

Add an `xml` file to the `public` dir and run a build. Visit
`localhost:3000/{newFileName}.xml` and see that it is served as
expected.