Fix: `CrossSellingShelf` Overrides Section v2 (#2168)
## What's the purpose of this pull request?

- https://github.com/vtex/faststore/pull/2165 introduces some bugs in
PDP while using `CrossSellingShelf` section. This PR aims to fix it.

This PR introduces `CrossSellingShelf` using Section Overrides v2.

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->

---------

Co-authored-by: Ícaro Azevedo <icaro.azevedo@vtex.com>