Stop appending `/` to search state base path (#2178)
## What's the purpose of this pull request?

When clicking on a suggested product link in the search the user would
be redirect to `/s/?term=blabla`. Because of how our nginx is
configured, this trailing `/` would redirect the user to the cluster URL
and would show a `Your connection is not private` error.

## How it works?

Stop appending `/` to the URL.

## How to test it?

Go to the preview link, start a search and click on a "Top Search"
suggestion. You should see the search page without any problems.