chore: improve overrides bundle size (#2186)
## What's the purpose of this pull request?

This PR improves the performance of the Section Override API.

## How it works?

It works by removing the centralized imports where an object would
contain the components for all of the sections. This made the sections
and default components to be imported even where they were not used,
making the bundle to be bigger than necessary.

It also introduces a breaking change to the previously proposed API:
instead of providing the name of the section to be overridden, users
should provide the section itself.

```tsx
import { Alert, getOverriddenSection } from "@faststore/core";

const NewAlert = getOverriddenSection({ Section: Alert, components: {} });

export default NewAlert;
```

## How to test it?

Use this version of the package on the starter and try to create
overrides for the supported V2 sections!