feat: Get CMS Pages by Version ID with fallback (#2215)
## What's the purpose of this pull request?

We need to stop getting the latest CMS page because of caching issues,
so we're going to retrieve the contents during the build time and using
those versions to request the necessary pages by version ID.

## How it works?

This PR uses the versionId to retrieve CMS page data, maintaining the
previous way of getting the latest version as fallback.

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->