fix: useQuery types & module resolution (#2241)
## What's the purpose of this pull request?

This PR fixes an issue where incompatible types between the return type
of the `gql` function and the `useQuery_unstable` experimental hook
would prevent the store from building.

It also fixes the issue where the module resolution for
`@faststore/core` exports would sometimes be referencing the exports
from the `.faststore` folder, which could lead to issues and inaccurate
types.

## How it works?

It works by better specifying the `useQuery` type for the operation, and
by removing exports from `.faststore/package.json` file. To compensate
for that, path aliases were added for `@faststore/core` exports in
`tsconfig.json`, so the `.faststore` uses its own definitions of the
exports when building.

## How to test it?

Define a query with `gql` and call it using `useQuery_unstable`. There
should be no errors. Then, build the store. There should be no errors.

### Starters Deploy Preview
https://github.com/vtex-sites/starter.store/pull/390