Fix SKU selector (#2254)
## What's the purpose of this pull request?

When implementing the Variant Swatch Image I noticed that if a SKU
Selector had only one option the image would not appear. Upon inspecting
this hook I noticed that if there's ony one image the code would infer
they were all the same and only show the label, but there's the special
case of only one image. This PR fixes this.

## How to test it?

Go to the preview and see the
[awesome-chair](https://sfj-5b2ff3a--starter.preview.vtex.app/awesome-chair-99988218/p)
product and switch to the Plastic or Iron material SKUs. You should see
an image on the Size selector. On the starter store
[awesome-chair](https://starter.vtex.app/awesome-chair-99988218/p) will
**not** show the image.