load environments variables in dev (#2276)
## What's the purpose of this pull request?

Env vars aren't available on development. So in this pr, I allow to load
`vtex.env` in the process.env in development.

## How it works?

All env var in `vtex.env` are available on `process.env` during
development.

## How to test it?

Create a custom env var in `vtex.env` and try to use it on your store.