Fix missing Checkbox props on CheckboxField component (#2281)
## What's the purpose of this pull request?

Our `CheckboxField` was inconsistent with its documentation. It was not
accepting the `checked`, `disabled`, and `partial` props from Checkbox.
This PR fixes this.

As a bonus, I've added Jest and React Testing Library so we can start
adding those tests to our components. I've first created a failing test
and then implemented the props.

## How to test it?

Just run `yarn test` and you should have all success specs.

### Starters Deploy Preview

For some reason we don't use the `CheckboxField` on our starter.

---------

Co-authored-by: eduardoformiga <eduardo.formiga@gmail.com>