Fetch multiple CMS template pages instead of just the first (#2314)
## What's the purpose of this pull request?

Currently, when trying to load contentTypes from the Headless CMS we
just fetch the first page and ignore the hasNextPage response from the
cmsClient.

## How it works?

This PR changes that to load all publishedContentTypes before returning.
This might be troublesome if a given account has thousands of content
types of the same type (PLP, for instance). We can either increase the
page size to make fewer requests or investigate if we can move the
filtering to the CMS Client.

## How to test it?

I've added tests, but this can be replicated by copying the
`server/cms/index.ts` code to a store's node_modules folder and you
should see the effect.