Stop running commonjs build in the background
Bash's `&` operator runs what's to the left of it on the background.
Usually it's used together with the `wait` command. This was not the
case. I **believe** that this command running in the background was what
was causing the intermitent issue we had. Whenever the esnext build
finished before the commonjs one, the process would end and kill the
background process.