Prepare for CLI to be run on any arbitrary path (#2368)
## What's the purpose of this pull request?

This PR still should not affect any behavior. Everything should work the
same as it did before.

To prepare the FastStore CLI to run on any arbitrary path. I've created
a closure called `withBasePath` that will receive the basePath and
export the same values as before. To keep the same behavior, wherever
the `directory.ts` file was being imported, I'll call the `withBasePath`
with `process.cwd()`

Starter PR 🔜