fix: Pass Through PLP URL Params (#2355)
## What's the purpose of this pull request?

URL params are stripped out on inbound PLP links when the app rebuilds
the URL from state. This is undesirable for end users who need to add
params to their links - e.g. referral or UTM tracking codes.

## How it works?

Any params that are not included in the state allow list are grouped
into a `passThrough` field in the state, and added back when the URL is
rebuilt.

## How to test it?

1. Visit a PLP link with a URL param: `/my-cool-category?foo=bar`
2. Validate that the additional param is preserved after URL state
updates

I'm happy to adjust naming or implementation, this set of changes was
the result of a spike into preserving UTM params on inbound links.

---------

Co-authored-by: Fanny Chien <fannychienn@gmail.com>