chore: update OrderSummary and ProductCard props description (#2383)
## What's the purpose of this pull request?

This PR intends to update the props descriptions for
[OrderSummary](https://developers.vtex.com/docs/guides/faststore/molecules-order-summary)
and
[ProductCard](https://developers.vtex.com/docs/guides/faststore/molecules-product-card).

## How it works?

This PR fixes some typos in the `OrderSummary` and `ProductCard` props
description. However, the main fixes are related to the taxes related
props as you can see in the following screenshot:

<img width="366" alt="Screenshot_578"
src="https://github.com/user-attachments/assets/72242278-14d4-48f4-ba85-de033e005b2e">


## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview


https://vercel.live/open-feedback/faststore-site-git-chore-update-props-description-faststore.vercel.app?via=pr-comment-visit-preview-link&passThrough=1

## References

Dev Portal PR (updating the documentation):
https://github.com/vtexdocs/dev-portal-content/pull/1298#pullrequestreview-2178476553